WinCLIP/requirements.txt
einops==0.8.0
filelock==3.13.1
fsspec==2024.2.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
imageio==2.36.0
iopath==0.1.10
Jinja2==3.1.3
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.10.0.84
packaging==24.1
pillow==11.0.0
portalocker==2.10.1
psutil==6.1.0
PyWavelets==1.7.0
PyYAML==6.0.2
regex==2024.9.11
scikit-image==0.21.0
scikit-learn==1.5.2
scipy==1.10.1
simplejson==3.19.3
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.64.0
triton==3.1.0
typing_extensions==4.9.0
wcwidth==0.2.13
yacs==0.1.8
